pyproject.toml
[tool.poetry]
name = "globoticket"
version = "0.1.0"
description = "Building a simple REST API with Python 3'"
authors = ["Juan Daniel Galindo <contact@gdanilo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.20"
python-dotenv = "^1.1.0"
pyyaml = "^6.0.2"
httpx = "^0.28.1"
uvicorn = "^0.34.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disable_error_code = ["import-untyped"]
